requirements.txt
sentence-transformers==2.2.2
pinecone-client==2.2.1
langchain==0.0.225
llama-cpp-python
flask
pypdf
python-dotenv
-e .